# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ALEF DELTA ERP API",
        "description": "Production-ready ERP API for manpower supply management with GPS-verified attendance, Telegram integration, payroll automation, and asset tracking.",
        "contact": {
            "email": "admin@alefdelta.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.alefdelta.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "List all employees",
                "operationId": "9aeb15c3a8c5b0b8b9e0931d1673e787",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ACTIVE",
                                "TERMINATED",
                                "ON_LEAVE"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of employees"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Create a new employee",
                "operationId": "2475af536398d342aca2fd7f60f9af3a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_code",
                                    "first_name",
                                    "last_name",
                                    "phone_number",
                                    "job_role_id",
                                    "hire_date"
                                ],
                                "properties": {
                                    "employee_code": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "job_role_id": {
                                        "type": "integer"
                                    },
                                    "hire_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "guarantor_details": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employees/{id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get employee details",
                "operationId": "41cfd4a7b54d2f35c31ad2afd9d37f3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Update employee details",
                "operationId": "f1e833660cfbccbfbdef00c84e24c5cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "employee_code": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "job_role_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "ACTIVE",
                                            "TERMINATED",
                                            "ON_LEAVE"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employee updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employees/link-telegram": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Link employee to Telegram account",
                "operationId": "bc16a13e54f34a35b52abcceaa6914ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_id",
                                    "telegram_chat_id"
                                ],
                                "properties": {
                                    "employee_id": {
                                        "type": "integer"
                                    },
                                    "telegram_chat_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Telegram linked successfully"
                    },
                    "404": {
                        "description": "Telegram user not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token obtained from login endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}